# swagger/v1/swagger.yaml
swagger: '2.0'
info:
  description: >-
    SMX api description.
  version: 1.0.0
  title: MoneyApi
  termsOfService: ''
  contact:
    email: 
host: 107.22.146.137
basePath: /api/v1
...
schemes:
  - http
securityDefinitions:
  X-API-APIKEY:
    name: X-API-APIKEY
    in: header
    description: logged in user api key
    type: apiKey
paths:
  /activities:
    get:
      tags:
        - activity
      summary: Get the list of user activities
      description: The endpoint will get the list of user activities in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      parameters:
        - in: query
          name: auth_token
          description: Auth token of the user
          required: true
          schema:
            $ref: '#/definitions/AuthToken'
        - in: query
          name: offset
          type: integer
        - in: query
          name: count
          type: integer
      responses:
        '200':
          description: List of user activities
          schema:
            $ref: '#/definitions/Activities'
        '401':
          $ref: '#/responses/401'
        '500':
          $ref: '#/responses/500'
...
  /notifications:
    get:
      tags:
        - notification
      summary: Get the list of user notifications
      description: The endpoint will get the list of user notifications in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      parameters:
        - in: query
          name: auth_token
          description: Auth token of the user
          required: true
          schema:
            $ref: '#/definitions/AuthToken'
        - in: query
          name: offset
          type: integer
        - in: query
          name: count
          type: integer
      responses:
        '200':
          description: List of user notifications
          schema:
            $ref: '#/definitions/Notifications'
        '401':
          $ref: '#/responses/401'
        '500':
          $ref: '#/responses/500'
  /notifications/update_notification:
    put:
      tags:
        - notification
      summary: Update user notification
      description: The endpoint will update user notification in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      parameters:
        - in: body
          name: body
          schema:
            allOf:
              - $ref: '#/definitions/AuthToken'
              - type: object
                properties:
                  notification:
                    properties:
                      id:
                        type: integer
                        example: 299
                      notice_type:
                        type: string
                        example: "TransferNotification"
                      option:
                        type: string
                        example: "Pay"
      responses:
        '201':
          description: Your transaction was successful
          schema:
            $ref: '#/definitions/SuccessResponse'
        '400':
          $ref: '#/responses/400'
        '401':
          $ref: '#/responses/401'
        '500':
          $ref: '#/responses/500'
  /notifications/notifications_count:
    get:
      tags:
        - notification
      summary: Get total count of unread notifications for the last 30 days
      description: The endpoint will get total count of unread notifications for the last 30 days in SMX system
      produces:
        - application/json
      consumes:
        - application/json
      security:
        - X-API-APIKEY: []
      parameters:
        - in: query
          name: auth_token
          description: Auth token of the user
          required: true
          schema:
            $ref: '#/definitions/AuthToken'
      responses:
        '200':
          schema:
            allOf:
              - type: object
                properties:
                  count:
                    type: integer
                    description: Notification count
                    example: 4
        '401':
          $ref: '#/responses/401'
        '404':
          $ref: '#/responses/404'
        '500':
          $ref: '#/responses/500'
...
definitions:
  AuthToken:
    type: object
    required:
      - auth_token
    properties:
      auth_token:
        type: string
        description: user api key
...
  ObjectID:
    type: object
    properties:
      id:
        type: integer
        description: Unique identifier
        example: 1
...
  Activity:
    properties:
      activity:
        type: string
        example: "Sent to david.bale2"
      amount:
        properties:
          value:
            type: string
            example: "50.0"
          currency:
            type: string
            example: "usd"
      status:
        type: string
        example: "Completed"
      date:
        type: string
        format: date-time
        example: '26 Jun 22:01'
  Activities:
    properties:
      activities:
        type: array
        items:
          allOf:
            - $ref: '#/definitions/ObjectID'
            - $ref: '#/definitions/Activity'
            - type: object
              properties:
                smx_transaction_id:
                  type: integer
                  example: 1
...
  Notifications:
    properties:
      notifications:
        type: array
        items:
          allOf:
            - $ref: '#/definitions/ObjectID'
            - type: object
              properties:
                notice:
                  type: string
                  example: "david.balea sent you 50.0"
                notice_type:
                  type: string
                  example: "TransferNotification"
                date:
                  type: string
                  format: date-time
                  example: '26 Jun 22:01'  
...
  Error:
    type: object
    required:
      - errors
    properties:
      errors:
        type: array
        items:
          type: string
          description: error message
  SuccessResponse:
    type: object
    required:
      - status
    properties:
      status:
        type: string
        description: Sucecess message
...
responses:
  '400':
    description: Input invalid or Bad Request
    schema:
      $ref: '#/definitions/Error'
  '401':
    description: Token Invlaid
    schema:
      $ref: '#/definitions/Error'
  '404':
    description: Resource not found
    schema:
      $ref: '#/definitions/Error'
  '500':
    description: Internal Server error
    schema:
      $ref: '#/definitions/Error'